pyproject.toml
[project]
name = "video-insertion"
version = "0.1.0"
description = "A modular pipeline for object-aware video insertion."
authors = [
    { name = "Your Name", email = "your@email.com" }
]
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    "torch",
    "torchvision",
    "opencv-python",
    "numpy",
    "matplotlib",
    "tqdm",
    "Pillow",
    "pyyaml",
    "scikit-image",
    "scikit-learn",
    "transformers",
    "sentence-transformers",
    "supervision>=0.26.1",
    "ultralytics>=8.3.170",
    "detectron2",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv.sources]
detectron2 = { git = "https://github.com/facebookresearch/detectron2.git" }


[project.optional-dependencies]
dev = [
    "black",
    "pytest",
    "isort",
    "pre-commit"
]

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"
